services/ms1-auth/requirements.txt
fastapi==0.115.7
pydantic==2.10.6
uvicorn==0.34.0
PyJWT==2.10.1
passlib==1.7.4
psycopg[binary]==3.2.6